Change all file format related unsigned longs to uint32_t,  so that 64 bits hosts can build this code correctly,  plus a little bit of Blackfin code.
Most of the changes provided Bernd Schmidt <bernds_cb1 at t-online.de>

+++ flat.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2002-2003  David McCullough <davidm@snapgear.com>
+ * Copyright (C) 2002-2005  David McCullough <davidm@snapgear.com>
  * Copyright (C) 1998       Kenneth Albanowski <kjahds@kjahds.com>
  *                          The Silver Hammer Group, Ltd.
  *
@@ -14,6 +14,7 @@
 #define _LINUX_FLAT_H
 
 #ifdef __KERNEL__
+#include <linux/types.h>
 #include <asm/flat.h>
 #endif
 
@@ -31,26 +32,26 @@
  */
 
 struct flat_hdr {
-	char magic[4];
-	unsigned long rev;          /* version (as above) */
-	unsigned long entry;        /* Offset of first executable instruction
-	                               with text segment from beginning of file */
-	unsigned long data_start;   /* Offset of data segment from beginning of
-	                               file */
-	unsigned long data_end;     /* Offset of end of data segment
-	                               from beginning of file */
-	unsigned long bss_end;      /* Offset of end of bss segment from beginning
-	                               of file */
+    char magic[4];
+    uint32_t rev;          /* version (as above) */
+    uint32_t entry;        /* Offset of first executable instruction
+                              with text segment from beginning of file */
+    uint32_t data_start;   /* Offset of data segment from beginning of
+                              file */
+    uint32_t data_end;     /* Offset of end of data segment from beginning
+                              of file */
+    uint32_t bss_end;      /* Offset of end of bss segment from beginning
+                              of file */
 
-	/* (It is assumed that data_end through bss_end forms the bss segment.) */
+    /* (It is assumed that data_end through bss_end forms the bss segment.) */
 
-	unsigned long stack_size;   /* Size of stack, in bytes */
-	unsigned long reloc_start;  /* Offset of relocation records from
-	                               beginning of file */
-	unsigned long reloc_count;  /* Number of relocation records */
-	unsigned long flags;       
-	unsigned long build_date;   /* When the program/library was built */
-	unsigned long filler[5];    /* Reservered, set to zero */
+    uint32_t stack_size;   /* Size of stack, in bytes */
+    uint32_t reloc_start;  /* Offset of relocation records from beginning
+                              of file */
+    uint32_t reloc_count;  /* Number of relocation records */
+    uint32_t flags;       
+    uint32_t build_date;   /* When the program/library was built */
+    uint32_t filler[5];    /* Reservered, set to zero */
 };
 
 #define FLAT_FLAG_RAM    0x0001 /* load program entirely into RAM */
@@ -58,7 +59,6 @@ struct flat_hdr {
 #define FLAT_FLAG_GZIP   0x0004 /* all but the header is compressed */
 #define FLAT_FLAG_GZDATA 0x0008 /* only data/relocs are compressed (for XIP) */
 #define FLAT_FLAG_KTRACE 0x0010 /* output useful kernel trace for debugging */
-
 
 #ifdef __KERNEL__ /* so systems without linux headers can compile the apps */
 /*
@@ -72,32 +72,40 @@ struct flat_hdr {
 
 #include <asm/byteorder.h>
 
-#define	OLD_FLAT_VERSION			0x00000002L
+#define	OLD_FLAT_VERSION		0x00000002L
 #define OLD_FLAT_RELOC_TYPE_TEXT	0
 #define OLD_FLAT_RELOC_TYPE_DATA	1
 #define OLD_FLAT_RELOC_TYPE_BSS		2
 
 typedef union {
-	unsigned long	value;
-	struct {
+    uint32_t        value;
+    struct {
 # if defined(mc68000) && !defined(CONFIG_COLDFIRE)
-		signed long offset : 30;
-		unsigned long type : 2;
+        int32_t     offset : 30;
+        uint32_t    type   : 2;
 #   	define OLD_FLAT_FLAG_RAM    0x1 /* load program entirely into RAM */
 # elif defined(__BIG_ENDIAN_BITFIELD)
-		unsigned long type : 2;
-		signed long offset : 30;
+        uint32_t    type   : 2;
+        int32_t     offset : 30;
 #   	define OLD_FLAT_FLAG_RAM    0x1 /* load program entirely into RAM */
 # elif defined(__LITTLE_ENDIAN_BITFIELD)
-		signed long offset : 30;
-		unsigned long type : 2;
+        int32_t     offset : 30;
+        uint32_t    type   : 2;
 #   	define OLD_FLAT_FLAG_RAM    0x1 /* load program entirely into RAM */
 # else
 #   	error "Unknown bitfield order for flat files."
 # endif
-	} reloc;
+    } reloc;
 } flat_v2_reloc_t;
 
 #endif /* __KERNEL__ */
 
 #endif /* _LINUX_FLAT_H */
+
+/* this __MUST__ be at the VERY end of the file - do NOT move!!
+ * Local Variables:
+ * c-basic-offset: 4
+ * tab-width: 8
+ * end:
+ * vi: tabstop=8 shiftwidth=4 textwidth=79 noexpandtab
+ */
